pyproject.toml
[tool.poetry]
name = "repl_python3_Boom"
version = "0.1.0"
description = ""
authors = ["IamEinstein <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
pymongo = "^3.11.4"
python-dotenv = "^0.18.0"
discord = "^1.7.3"
flask = "^2.0.1"
umongo = "^3.0.0"
marshmallow = "^3.12.1"
dnspython = "^2.1.0"
discord-py-interactions = "^3.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
